Bug fixes and verifying call-back for not null before calling
cases in test application where improper content service object is
reused.
Handled crash when settings modified and application relaunched for
modifying the settings, verifying if fragment is added to activity
before accessing the resource.
Handled EditTextPreference issue to display the wrong text/summary,
setting the text to store in preferences and using set summary to
display the summary
Verify callbacks are not null