Using Auto Play Channel Property
Auto play property of channel can be used to autoplay media in channel
playlist. Currently, autoplay value is always coming false. So cant use
this value as of now. When server starts sending right value for auto
play then uncomment the  code getting checked in this commit and do not
expose the setAutoPlay () API and also test app should remove the code
where it is setting autoplay for channel playlist.